requirements.txt
MetaTrader5==5.0.44
pandas==2.0.3
numpy==1.26.4
torchvision==0.20.1
torchaudio==2.5.1
scikit-learn==1.3.0
# ta-lib will be installed via prebuilt wheel manually on Windows
yfinance==0.2.18
requests==2.31.0
schedule==1.2.0
pyyaml==6.0
matplotlib==3.7.2
plotly==5.15.0
websocket-client==1.6.3
alpha-vantage==2.3.1
tensorflow==2.20.0
tensorboard==2.20.0
xgboost==1.7.6
lightgbm==4.1.0
catboost==1.2
